Added lfsr_mdir_sync to help sync mdir state
Syncing mdirs got a bit awkward with the addition of the gcksumdelta, we
can't just copy the rbyd component. So added lfsr_mdir_sync to abstract
out the copying of any relevent mdir state.

Note this does _not_ include the mid, which may be different for
different files in the same mdir.

The code deduplication ended up saving a bit of code, which is nice:

           code          stack          ctx
  before: 38668           2648          644
  after:  38612 (-0.1%)   2648 (+0.0%)  644 (+0.0%)