Delayed appending any gdelta until the last mdir commit
It's not the most intuitive, but we always know if an mdir commit will
be the last commit/atomic before attempting the commit. This means we
can wait to append/flush gdelta until the last commit, which simplifies
handling of gdelta state in lfsr_mdir_commit a bit.

We were already doing this for the gcksum, this just changes the grm to
match.

This also results in gdeltas being pushed upwards when relocating, which
is probably a good thing? On one hand this encourages deduplication of
gdeltas, on the other hand it does result in gdeltas getting pushed into
mroots that don't need them (grm in mrootanchor?).

This ends up saving a bit of code, which is nice:

           code          stack          ctx
  before: 38612           2648          644
  after:  38588 (-0.1%)   2648 (+0.0%)  644 (+0.0%)