Added LFS_LITTLER for disabling non-essential/debugging utils
This is a simple alias for a number of existing defines:

  LFS_LITTLER => LFS_NO_DEBUG
                 LFS_NO_INFO
                 LFS_NO_WARN
                 LFS_NO_ERROR
                 LFS_NO_ASSERT

This is mainly to make CI/code measurements easier. I suspect most
systems already make these zero (or at least minimal) cost in release
mode, but when building with our Makefile we need some way to disable
the simple printf/assert mapping.

This probably deserves a better name, but at least this is something.