Added LFS_BIGGEST for enabling all opt-in features
Like LFS_LITTLER, this is a simple alias for existing defines. In this
case it enables all opt-in features, which is useful for testing:

  LFS_BIGGEST => LFS_CKPROGS
                 LFS_CKFETCHES
                 LFS_CKPARITY
                 LFS_CKDATACKSUMS
                 LFS_GC

This lowers the risk of forgetting a feature when running tests, which
has already happend a couple times.

The name comes from the idea of adding other tiers of features (
LFS_LITTLER, LFS_LITTLERER, LFS_BIGGER, etc), but I'm not entirely sure
what these would all look like.