Fixed test_traversal_compact_mroot_split NOSPC error
This is a pretty suspicious looking test failure, considering the recent
changes to the mroot/mtree and related splitting logic, but it just
turned out to be a bug in the test logic.

Sort of. This loop is trying to create an mroot that will both compact
and split, but it doesn't check if the mdir was split prematurely, so it
just keeps adding files until we hit a true LFS_ERR_NOSPC condition:

  if ((file1.o.o.mdir.rbyd.eoff & 0x7fffffff) > GC_COMPACT_THRESH
          && estimate > BLOCK_SIZE/2) {
      break;
  }

The solution is to make the filename size a bit smaller so we don't
split too early.

I also added some asserts to catch premature splits in case this happens
again. These tests are a bit delicate.